Split URL+TemplateExtraction.swift into RFC6570 extension files
Goal: bring the file under file_length (1000) and the
RFC6570TemplateExtractor type body under type_body_length (350). Also
decompose two long extract* functions that exceeded
function_body_length / cyclomatic_complexity.

Before: 1268-line monolith with `private struct RFC6570TemplateExtractor`
(675-line body) and `private struct RFC6570TemplateConstructor`.
After: 106-line main file (URL/String public extensions + extractor
properties + main dispatch only), with 8 sibling files:

- URI+Construction.swift — RFC6570TemplateConstructor extracted.
- RFC6570TemplateExtractor+Types.swift — ExpressionOperator,
  VariableSpec, VariableModifier, ExtractedVariable.
- RFC6570TemplateExtractor+Parsing.swift — parseExpression,
  parseVariableSpec, parseQueryParameters, fragment/query expression
  extraction.
- RFC6570TemplateExtractor+Base.swift — extractFromBase (decomposed
  via a `BaseCursor` helper into consumeExpression / literalFollowing
  / finalizeBaseMatch) + dispatchers.
- RFC6570TemplateExtractor+Simple.swift — simple + reserved variants.
- RFC6570TemplateExtractor+Label.swift — label variant.
- RFC6570TemplateExtractor+Path.swift — path-segment + path-style,
  long functions split into consumePathSegment /
  collectExplodedPathSegments / consumePathStyleParameter /
  readPathStyleValue.
- RFC6570TemplateExtractor+Query.swift — query + query-continuation.

Access widening from private → internal: RFC6570TemplateExtractor,
RFC6570TemplateConstructor, ExpressionOperator, VariableSpec,
VariableModifier (file-scope `private` doesn't cross file boundaries).

Build clean. swift test --filter RFC6570ExtractionTests (29/29),
URIConstructionTests (10/10), full suite (415/415).
